feature:daily 모듈의 View 기반 코드 삭제 (#131)
* #130 - remove : daily 모듈의 View기반 코드 삭제

* #130 - style : daily모듈의 screen패키지에 있던것들을 screen패키지 상위 패키지로 이동
