+++ feature/daily/src/androidTest/java/com/hegunhee/daily/DailyScreenTest.kt
@@ -3,10 +3,8 @@ package com.hegunhee.daily
 import android.content.Context
 import androidx.activity.ComponentActivity
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsNotDisplayed
@@ -19,8 +17,6 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import com.hegunhee.daily.screen.DailyScreen
-import com.hegunhee.daily.screen.DailyUiState
 import hegunhee.routiner.model.Routine
 import org.junit.Before
 import org.junit.Rule
