--- feature/daily/src/main/java/com/hegunhee/daily/DailyActionHandler.kt
@@ -1,12 +1,0 @@
-package com.hegunhee.daily
-
-import hegunhee.routiner.model.Routine
-
-interface DailyActionHandler {
-
-    fun deleteRoutine(routineNum : Int)
-
-    fun toggleFinishRoutine(routine : Routine)
-
-    fun onClickRoutineInsert()
-}
