--- feature/daily/src/main/java/com/hegunhee/daily/DailyNavigationAction.kt
@@ -1,6 +1,0 @@
-package com.hegunhee.daily
-
-sealed interface DailyNavigationAction {
-
-    object InsertRoutine : DailyNavigationAction
-}
