+++ feature/daily/src/main/java/com/hegunhee/daily/DailyViewModel.kt
@@ -1,17 +1,16 @@
 package com.hegunhee.daily
 
-import androidx.lifecycle.*
-import hegunhee.routiner.model.Routine
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.example.domain.usecase.routine.DeleteRoutineUseCase
 import com.example.domain.usecase.routine.GetRoutinesFlowByDateUseCase
 import com.example.domain.usecase.routine.UpdateRoutineUseCase
+import com.hegunhee.daily.DailyUiState.Init
 import com.hegunhee.routiner.util.getTodayDate
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.SharedFlow
+import hegunhee.routiner.model.Routine
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
@@ -19,45 +18,35 @@ import javax.inject.Inject
 
 @HiltViewModel
 class DailyViewModel @Inject constructor(
-    private val getRoutinesFlowByDateUseCase: GetRoutinesFlowByDateUseCase,
+    private val getRoutinesByFlow: GetRoutinesFlowByDateUseCase,
+    private val updateRoutineUseCase: UpdateRoutineUseCase,
     private val deleteRoutineUseCase: DeleteRoutineUseCase,
-    private val updateRoutineUseCase: UpdateRoutineUseCase
-) : ViewModel(), DailyActionHandler {
+) : ViewModel() {
 
-    val dailyRoutineEntityList: StateFlow<List<Routine>> = getRoutinesFlowByDateUseCase(
-        getTodayDate()
-    ).stateIn(
-        scope = viewModelScope,
-        started = SharingStarted.WhileSubscribed(500L),
-        initialValue = emptyList()
-    )
+    private val routines = getRoutinesByFlow(getTodayDate())
 
-    val isDailyRoutineEmpty: StateFlow<Boolean> = dailyRoutineEntityList.map{
-        dailyRoutineEntityList.value.isEmpty()
+    val uiState : StateFlow<DailyUiState> = routines.map {
+        if(it.isEmpty()) {
+            DailyUiState.Empty
+        } else {
+            DailyUiState.Items(it)
+        }
     }.stateIn(
         scope = viewModelScope,
         started = SharingStarted.WhileSubscribed(500L),
-        initialValue = true
+        initialValue = Init
     )
 
-    private val _navigationActions : MutableSharedFlow<DailyNavigationAction> = MutableSharedFlow()
-    val navigationActions : SharedFlow<DailyNavigationAction> = _navigationActions.asSharedFlow()
-
-    override fun onClickRoutineInsert() {
-        viewModelScope.launch{
-            _navigationActions.emit(DailyNavigationAction.InsertRoutine)
-        }
-    }
-
-    override fun deleteRoutine(routineNum: Int) {
-        viewModelScope.launch{
-            deleteRoutineUseCase(routineNum)
-        }
-    }
-
-    override fun toggleFinishRoutine(routine: Routine) {
+    fun onClickDailyRoutine(routine: Routine) {
         viewModelScope.launch {
             updateRoutineUseCase(routine.copy(isFinished = !routine.isFinished))
         }
     }
+
+    fun onClickDeleteDailyRoutine(routineId: Int) {
+        viewModelScope.launch {
+            deleteRoutineUseCase(routineId)
+        }
+    }
+
 }
