--- feature/daily/src/main/res/layout/fragment_daily.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-
-    <data>
-
-        <variable
-            name="viewmodel"
-            type="com.hegunhee.daily.DailyViewModel" />
-
-        <import type="android.view.View" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/daily_recyclerView"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:visibility="@{viewmodel.isDailyRoutineEmpty ? View.GONE : View.VISIBLE}"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/empty_view"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:visibility="@{viewmodel.isDailyRoutineEmpty ? View.VISIBLE : View.GONE}"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <TextView
-                android:id="@+id/empty_textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center"
-                android:text="오늘 등록된 루틴이 없습니다.\n 루틴을 등록해주세요"
-                android:textColor="@color/white"
-                android:textSize="30sp"
-                android:textStyle="bold"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <Button
-                android:id="@+id/empty_addButton"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="20dp"
-                android:onClick="@{() -> viewmodel.onClickRoutineInsert()}"
-                android:text="루틴 등록하기"
-                app:layout_constraintEnd_toEndOf="@id/empty_textView"
-                app:layout_constraintStart_toStartOf="@id/empty_textView"
-                app:layout_constraintTop_toBottomOf="@id/empty_textView" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/floating_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginVertical="10dp"
-            android:layout_marginEnd="10dp"
-            android:layout_marginBottom="100dp"
-            android:backgroundTint="@color/sky_blue"
-            android:contentDescription="@string/add_routine"
-            android:onClick="@{() -> viewmodel.onClickRoutineInsert()}"
-            android:src="@drawable/ic_add"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
