--- feature/daily/src/main/res/layout/item_daily.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-
-    <data>
-
-        <variable
-            name="eventHandler"
-            type="com.hegunhee.daily.DailyActionHandler" />
-
-        <variable
-            name="routine"
-            type="hegunhee.routiner.model.Routine" />
-
-        <import type="android.view.View" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@drawable/rounded_shape_gray_20dp"
-        android:layout_marginHorizontal="10dp"
-        android:backgroundTint="@color/gray"
-        android:layout_marginVertical="10dp"
-        android:onClick="@{() -> eventHandler.toggleFinishRoutine(routine)}"
-        android:paddingHorizontal="10dp">
-
-        <CheckBox
-            android:id="@+id/checkBox"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:onClick="@{() -> eventHandler.toggleFinishRoutine(routine)}"
-            android:checked="@{routine.finished ? true : false}"/>
-
-        <TextView
-            android:id="@+id/routine_title"
-            android:layout_width="150dp"
-            android:layout_height="48dp"
-            android:layout_marginStart="10dp"
-            android:layout_marginTop="10dp"
-            android:layout_marginBottom="10dp"
-            android:ellipsize="end"
-            android:maxLines="1"
-            android:text="@{routine.text}"
-            android:textSize="30sp"
-            android:textColor="@color/white"
-            app:layout_constraintStart_toEndOf="@id/checkBox"
-            app:layout_constraintTop_toTopOf="@id/checkBox"
-            app:layout_constraintBottom_toBottomOf="@id/checkBox"
-            tools:text="영어 공부하기" />
-
-        <TextView
-            android:id="@+id/category"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toBottomOf="@id/routine_title"
-            app:layout_constraintStart_toStartOf="@id/routine_title"
-            android:visibility="@{routine.category.empty ? View.GONE : View.VISIBLE}"
-            android:text="@{routine.category}"
-            android:textSize="15sp"
-            android:textColor="@color/white"
-            tools:text="공부"
-            android:ellipsize="end"
-            android:layout_marginStart="10dp"
-            android:paddingBottom="5dp"/>
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/deleteButton"
-            android:layout_width="wrap_content"
-            android:layout_height="48dp"
-            android:background="@drawable/rounded_shape_gray_20dp"
-            android:backgroundTint="@color/sky_blue"
-            android:elevation="5dp"
-            android:onClick="@{() -> eventHandler.deleteRoutine(routine.id)}"
-            android:text="Delete"
-            android:textColor="@color/black"
-            app:layout_constraintBottom_toBottomOf="@id/checkBox"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/checkBox" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
