--- feature/daily/src/main/res/layout/item_daily_header.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/dailyTitle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="루틴 리스트"
-            android:textColor="@color/white"
-            android:textStyle="bold"
-            android:textSize="30sp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginTop="5dp"
-            android:layout_marginStart="5dp"/>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
