+++ feature/daily/src/main/res/values/strings.xml
@@ -4,7 +4,6 @@
     <string name="empty_string">오늘 등록된 루틴이 없습니다.\n 루틴을 등록해주세요</string>
     <string name="add_routine_string">루틴 등록하기</string>
     <string name="delete_routine">루틴 삭제</string>
-    <string name="add_routine">루틴 추가</string>
     
     <string name="routine_list_text">루틴 리스트</string>
     <string name="routine_degree">작업 진척도</string>
